Refactor the routes logic
